requirements.txt
fastapi
uvicorn
python-dotenv
opengradient
